+++ cli/tests/pcluster/models/imagebuilder_dummy_model.py
@@ -9,18 +9,18 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 from pcluster.models.common import BaseTag, Cookbook
-from pcluster.models.imagebuilder import (
+from pcluster.models.imagebuilder_config import (
     Build,
     Component,
     DistributionConfiguration,
     Image,
-    ImageBuilder,
+    ImageBuilderConfig,
     ImagebuilderDevSettings,
     Volume,
 )
 
 CLASS_DICT = {
-    "imagebuilder": ImageBuilder,
+    "imagebuilder": ImageBuilderConfig,
     "image": Image,
     "build": Build,
     "dev_settings": ImagebuilderDevSettings,